integrated react router dom with basic cart component and improvement
