+++ src/components/cart/Cart.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Cart = () => {
+  return <div>THis is cart</div>;
+};
+
+export default Cart;
