+++ src/components/header/Header.js
@@ -1,12 +1,14 @@
 import React from "react";
+import { Link } from "react-router-dom";
+import './header.scss';
 
 const Header = React.memo((props) => {
-    return (
-      <header className="header">
-        <div>My App</div>
-        <div>Cart ({props.cartCount})</div>
-      </header>
-    );
+  return (
+    <header className="header">
+      <Link className="header-title" to="/">ShopKart</Link>
+      <Link className="header-cart-count" to="/cart">Cart ({props.cartCount})</Link>
+    </header>
+  );
 });
 
 export default Header;
