+++ src/components/header/header.scss
@@ -1,0 +1,13 @@
+.header {
+    background-color: black;
+    color: white;
+    display: flex;
+    padding: 20px 50px;
+    justify-content: space-between;
+    text-decoration: none;
+
+    &-title, &-cart-count {
+        text-decoration: none;
+        color: white;
+    }
+  }
