+++ src/components/home/home.scss
@@ -1,0 +1,14 @@
+.home-header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
+
+  .home {
+    text-align: center;
+  }
